fix db bag
я виправив той баг і додав виведення всіх юзерів, але на базовому рівні
