+++ db_requsts/user.py
@@ -1,0 +1,8 @@
+from models import (
+    User,
+    Session
+)
+
+def  get_all_users():
+    with Session() as session:
+        return session.query(User).all()
