+++ models/user.py
@@ -3,7 +3,6 @@ from sqlalchemy import String, ForeignKey
 from typing import Optional, List
 
 from .base import Base
-from .role import Role
 from .associates import user_cart_assoc_table, user_favorites_assoc_table
 from .product import Product
 
@@ -17,8 +16,8 @@ class User(Base):
     last_name: Mapped[Optional[str]] = mapped_column(String(50))
     phone_number: Mapped[Optional[str]] = mapped_column(String(15))
     zip_code: Mapped[Optional[str]] = mapped_column(String(50))
-    role_id: Mapped[Optional[int]] = ForeignKey('role.id', ondelete='CASCADE')
+    role_id: Mapped[Optional[int]] = mapped_column(ForeignKey('role.id', ondelete='CASCADE'))
 
-    role: Mapped[Optional[Role]] = relationship(back_populates="user")
+    role: Mapped[Optional["Role"]] = relationship(back_populates="users")
     cart: Mapped[List[Product]] = relationship(secondary=user_cart_assoc_table, back_populates="cart_users")
     favorites: Mapped[List[Product]] = relationship(secondary=user_favorites_assoc_table, back_populates="fav_users")
