+++ routes/__init__.py
@@ -1,0 +1,2 @@
+from routes.user import *
+from routes.default import *
